# SNAP-2931 - counting each event type separately and submitting only