those operations which are required

## Changes proposed in this pull request

Earlier we were counting only delete and update events and skipping both the operations if no update or delete events exists. Insert job was submitted irrespective of the presence of insert events.

With this change, we are counting events for all event types and performing the operation only if respective events are present in the incoming batch.

## Patch testing

Manual testing, Existing sink tests passing, Some performance tests

Following are the readings for the time taken by count job for older and newer approach:

### new approach:

- 50k 2 seconds
- 100k 2 seconds
- 1m 11 seconds

### old approach:

- 50k 1 second
- 100k 5 seconds
- 1m 12 seconds